requirements.txt
torchvision
unidecode
numpy
tqdm
tensorboard
matplotlib

pandas
datasets
librosa
wandb